Preventing deletion of Product if the same is related to some Attendance. Excluded substration of quantity in stock and used quantity of Product

+++ projeto-final-paola180115-fernando180016/src/main/java/com/facens/controller/ProductController.java
@@ -82,6 +82,10 @@ public class ProductController {
 	public String delete (@RequestParam Integer id)
 	{
 		try {
+			if (ps.getUsedProductsCountByProductId (id) != 0) {
+				return "redirect:/errorPage";
+			}
+
 			Product p = ps.getProductById (id);
 			if (p != null) {
 				ps.deleteById (id);
